local-ai/tts-inference-server/requirements.txt
# TTS Inference Server Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Chatterbox TTS
chatterbox-tts>=0.1.0

# PyTorch with CUDA (pre-installed in base image, but specify for completeness)
# These are installed via pip with --index-url for CUDA support
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
soundfile>=0.12.0
